feat: separate InitError into different file #447
