+++ PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/api/exceptions/InitError.java
@@ -1,0 +1,15 @@
+package org.prebid.mobile.api.exceptions;
+
+public class InitError {
+
+    private String error;
+
+    public InitError(String error) {
+        this.error = error;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+}
