+++ PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/rendering/listeners/SdkInitializationListener.java
@@ -1,23 +1,11 @@
 package org.prebid.mobile.rendering.listeners;
 
+import org.prebid.mobile.api.exceptions.InitError;
+
 public interface SdkInitializationListener {
 
     void onSdkInit();
 
     void onSdkFailedToInit(InitError error);
-
-    class InitError {
-
-        private String error;
-
-        public InitError(String error) {
-            this.error = error;
-        }
-
-        public String getError() {
-            return error;
-        }
-
-    }
 
 }
